requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
streamlit==1.28.2
python-dotenv==1.0.0
openai==1.3.5
pytest==7.4.3
black==23.11.0
isort==5.12.0
flake8==6.1.0
webdriver-manager==4.0.1 